% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dem.R
\name{bem_data}
\alias{bem_data}
\title{Download and load digital elevation data of Berlin}
\usage{
bem_data(
  x,
  directory = "BEMdata",
  read = TRUE,
  output_format = "raster",
  merge_output = TRUE
)
}
\arguments{
\item{x}{The target area for which dem data are to be downloaded.
Either created with \code{bem_target} or any object which can be 
passed to \code{sf::st_bbox}}

\item{directory}{A string giving the path to the directory where downloaded 
.zip files should be saved to}

\item{read}{Logical. Should the downloaded data be read into R?}

\item{output_format}{Character specifying the output format.
Default "raster" will return output as \code{raster::`Raster-class`}.
 The alternative is "tbl" which will return a 
 \code{tibble::`tibble-package`} of xyz coordinates instead.}

\item{merge_output}{Logical. If \code{TRUE}, all downloaded tiles will be 
merged in a single object. If \code{FALSE}, a list of objects with one entry
per tile will be returned}
}
\value{
Either a \code{raster::`Raster-class`}, a 
\code{tibble::`tibble-package`} or a \code{list} of any of those things,
depending on \code{output_format}. If \code{read = FALSE}, \code{x} 
is returned invisible.
}
\description{
Download and load digital elevation data of Berlin
}
\details{
This function will identify targeted tiles of the 1m resolution 
digital elevation model of Berlin that are distributed by the
\href{https://fbinter.stadt-berlin.de/fb/atom//Blattschnitte/2X2_EPSG_25833.gif}{Geoportal Berlin}
("Geoportal Berlin / ATKRIS(R) DGM (1m-Rasterweite)" \url{https://www.govdata.de/dl-de/by-2-0})
and download them into a specified directory (default is "BEMdata/"). It
will then attempt to load all downloaded data into R and return it in a 
meaningful format that should make further processing easy.
}
\examples{
\dontrun{
my_area_of_interest <- bem_target(
xmin = 384000, 
xmax = 386000, 
ymin = 5806000, 
ymax = 5808000)

# Careful! This will kick of ~ 100 MB download!
result <- bem_data(my_area_of_interest)

# In this case the result can be examined e.g. with the help of

# raster::plot(result)

# or

# mapview::mapview(result)
     
}
  
  

}
