% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{bem_select_fields}
\alias{bem_select_fields}
\title{Convert target geometry to download links}
\usage{
bem_select_fields(x)
}
\arguments{
\item{x}{A \code{sf::st_bbox} or any object coercible to one}
}
\value{
a [sf][sf::`sf-package`]
}
\description{
Convert target geometry to download links
}
\details{
EPSG: 25833
}
