% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dem.R
\name{bem_target}
\alias{bem_target}
\title{Specify a rectangular target area for dem data}
\usage{
bem_target(xmin, xmax, ymin, ymax)
}
\arguments{
\item{xmin}{The left boundary of the target rectangle}

\item{xmax}{The right boundary of the target rectangle}

\item{ymin}{The lower boundary of the target rectangle}

\item{ymax}{The upper boundary of the target rectangle}
}
\value{
A \code{\link[sf:st_bbox]{Bounding Box}}
}
\description{
Specify a rectangular target area for dem data
}
\details{
EPSG: 25833
}
\examples{
bem_target(
xmin = 384000, 
xmax = 386000, 
ymin = 5806000, 
ymax = 5808000)
 
}
