% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem_Index}
\alias{dem_Index}
\title{DEM tiles available for download}
\format{A \code{\link[sf:sf]{simple feature}}, a data.frame like table 
structure containing geographic information}
\source{
\href{https://fbinter.stadt-berlin.de/fb/atom//Blattschnitte/2X2_EPSG_25833.gif}{Geoportal Berlin}
}
\usage{
dem_Index
}
\description{
A vector-geodataset containing the positions, names and the extend of tiles of the 
digital elevation model of Berlin which can be accessed via this package
}
\details{
EPSG: 25833
licence: "Geoportal Berlin / ATKRIS(R) DGM (1m-Rasterweite)" \url{https://www.govdata.de/dl-de/by-2-0}
}
\keyword{datasets}
