% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{districts}
\alias{districts}
\title{City Districts of Berlin}
\format{A \code{\link[sf:sf]{simple feature}}, a data.frame like table 
structure containing geographic information}
\source{
\href{`https://fbinter.stadt-berlin.de/fb/berlin/service_intern.jsp?id=re_ortsteil@senstadt&type=WFS&themeType=spatial`}{Geoportal Berlin}
}
\usage{
districts
}
\description{
A vector-geodataset of polygons of the city districts of Berlin
}
\details{
EPSG: 25833
licence: "Geoportal Berlin / Ortsteile" \url{https://www.govdata.de/dl-de/by-2-0}
}
\keyword{datasets}
